correccion en migraciones

+++ database/migrations/2025_11_29_061156_create_chat_mensajes_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id();
 
             // Voluntario / usuario participante del chat
-            $table->unsignedInteger('voluntario_id');
+            $table->unsignedBigInteger('voluntario_id');
 
             // Quién envía el mensaje
             $table->enum('de', ['voluntario', 'admin']);
